hospital-microservices/auth-service/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
psycopg2-binary==2.9.7
drf-spectacular==0.26.5
pika==1.3.2
python-dotenv==1.0.0
cryptography==41.0.7
requests==2.31.0
